Removed code supporting old JDK
- We already use release 11
- Removed code using Unsafe which is not supported on JDK23 any more


# Conflicts:
#	pom.xml
